fix: move currencies to currency model
